refactoring distributions again
