+++ src/exoplanet/distributions/eccentricity.py
@@ -1,86 +1,10 @@
-__all__ = [
-    "angle",
-    "unit_disk",
-    "quad_limb_dark",
-    "impact_parameter",
-    "kipping13",
-    "vaneylen19",
-]
+__all__ = ["kipping13", "vaneylen19"]
 
 import numpy as np
 
 from exoplanet.citations import add_citations_to_model
 from exoplanet.compat import USING_PYMC3, pm
 from exoplanet.compat import tensor as at
-from exoplanet.utils import as_tensor_variable
-
-
-def with_initval(val, **kwargs):
-    if USING_PYMC3:
-        return dict(kwargs, testval=val)
-    return dict(kwargs, initval=val)
-
-
-def angle(name, *, regularization=10.0, **kwargs):
-    initval = kwargs.pop("initval", kwargs.pop("testval", 0.0))
-    x1 = pm.Normal(
-        f"{name}_angle1__", **with_initval(np.sin(initval), **kwargs)
-    )
-    x2 = pm.Normal(
-        f"{name}_angle2__", **with_initval(np.cos(initval), **kwargs)
-    )
-    if regularization is not None:
-        pm.Potential(
-            f"_{name}_regularization",
-            regularization * at.log(x1**2 + x2**2),
-        )
-    return pm.Deterministic(name, at.arctan2(x1, x2))
-
-
-def unit_disk(name_x, name_y, **kwargs):
-    initval = kwargs.pop("initval", kwargs.pop("testval", [0.0, 0.0]))
-    kwargs["lower"] = -1.0
-    kwargs["upper"] = 1.0
-    x1 = pm.Uniform(name_x, **with_initval(initval[0], **kwargs))
-    x2 = pm.Uniform(
-        f"{name_y}_unit_disk__",
-        **with_initval(initval[1] * np.sqrt(1 - initval[0] ** 2), **kwargs),
-    )
-    norm = at.sqrt(1 - x1**2)
-    pm.Potential(f"{name_y}_jacobian", at.log(norm))
-    return x1, pm.Deterministic(name_y, x2 * norm)
-
-
-def quad_limb_dark(name, **kwargs):
-    u = kwargs.pop("initval", kwargs.pop("testval", [np.sqrt(0.5), 0.0]))
-    u1 = u[0]
-    u2 = u[1]
-    kwargs["lower"] = 0.0
-    kwargs["upper"] = 1.0
-    q1 = pm.Uniform(f"{name}_q1__", **with_initval((u1 + u2) ** 2, **kwargs))
-    q2 = pm.Uniform(
-        f"{name}_q2__", **with_initval(0.5 * u1 / (u1 + u2), **kwargs)
-    )
-    sqrtq1 = at.sqrt(q1)
-    twoq2 = 2 * q2
-    return pm.Deterministic(
-        name, at.stack([sqrtq1 * twoq2, sqrtq1 * (1 - twoq2)], axis=0)
-    )
-
-
-def impact_parameter(name, ror, **kwargs):
-    ror = as_tensor_variable(ror)
-    bhat = kwargs.pop("initval", kwargs.pop("testval", 0.5))
-    if not USING_PYMC3:
-        shape = kwargs.setdefault("shape", ror.shape)
-        bhat = at.broadcast_to(bhat, shape)
-    kwargs["lower"] = 0.0
-    kwargs["upper"] = 1.0
-    norm = pm.Uniform(
-        f"{name}_impact_parameter__",
-        **with_initval(bhat / (1 + ror), **kwargs),
-    )
-    return pm.Deterministic(name, norm * (1 + ror))
 
 
 def kipping13(
@@ -146,7 +70,7 @@ def kipping13(
                     lower=0,
                     mu=alpha_mu,
                     sigma=alpha_sd,
-                    **with_initval(alpha_mu),
+                    **_with_initval(alpha_mu),
                 )
                 beta = _truncate(
                     "beta",
@@ -154,7 +78,7 @@ def kipping13(
                     lower=0,
                     mu=beta_mu,
                     sigma=beta_sd,
-                    **with_initval(beta_mu),
+                    **_with_initval(beta_mu),
                 )
 
         # Allow for upper and lower bounds
@@ -240,7 +164,7 @@ def vaneylen19(
                     lower=0,
                     mu=sigma_gauss_mu,
                     sigma=sigma_gauss_sd,
-                    **with_initval(sigma_gauss_mu),
+                    **_with_initval(sigma_gauss_mu),
                 )
                 sigma_rayleigh = _truncate(
                     "sigma_rayleigh",
@@ -248,7 +172,7 @@ def vaneylen19(
                     lower=0,
                     mu=sigma_rayleigh_mu,
                     sigma=sigma_rayleigh_sd,
-                    **with_initval(sigma_rayleigh_mu),
+                    **_with_initval(sigma_rayleigh_mu),
                 )
                 frac = _truncate(
                     "frac",
@@ -257,7 +181,7 @@ def vaneylen19(
                     upper=1,
                     mu=frac_mu,
                     sigma=frac_sd,
-                    **with_initval(frac_mu),
+                    **_with_initval(frac_mu),
                 )
 
             gauss = pm.HalfNormal.dist(sigma=sigma_gauss)
@@ -274,6 +198,12 @@ def vaneylen19(
             )
 
         return ecc
+
+
+def _with_initval(val, **kwargs):
+    if USING_PYMC3:
+        return dict(kwargs, testval=val)
+    return dict(kwargs, initval=val)
 
 
 def _logp(rv, x):
