+++ tests/distributions_test.py
@@ -5,12 +5,11 @@ import pytest
 from scipy.stats import beta, halfnorm, kstest, rayleigh
 
 from exoplanet.compat import USING_PYMC3, pm
-from exoplanet.distributions import (
+from exoplanet.distributions.distributions import (
     impact_parameter,
-    kipping13,
     quad_limb_dark,
-    vaneylen19,
 )
+from exoplanet.distributions.eccentricity import kipping13, vaneylen19
 
 
 class _Base:
